replaced texture type with texture usage

+++ library/graphics_api/src/TimestampArray.cpp
@@ -4,7 +4,7 @@
 
 namespace triglav::graphics_api {
 
-TimestampArray::TimestampArray(vulkan::QueryPool queryPool, const u64 timestampPeriod) :
+TimestampArray::TimestampArray(vulkan::QueryPool queryPool, const float timestampPeriod) :
     m_queryPool(std::move(queryPool)),
     m_timestampPeriod(timestampPeriod)
 {
@@ -15,7 +15,7 @@ void TimestampArray::get_result(std::span<u64> out, const u32 first) const
    vkGetQueryPoolResults(m_queryPool.parent(), *m_queryPool, first, out.size(), sizeof(u64) * out.size(),
                          out.data(), sizeof(u64), VK_QUERY_RESULT_64_BIT | VK_QUERY_RESULT_WAIT_BIT);
 
-   for (auto& value : out) {
+   for (auto &value : out) {
       value *= m_timestampPeriod;
    }
 }
